requirements.txt
# Patient Safety Monitor - Python Dependencies
# Version: 1.0.0

# =============================================================================
# Core Framework
# =============================================================================
pydantic>=2.5.0
pydantic-settings>=2.1.0

# =============================================================================
# Database
# =============================================================================
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# =============================================================================
# Web Scraping
# =============================================================================
httpx>=0.26.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
playwright>=1.40.0

# PDF Processing
pdfplumber>=0.10.0
pypdf>=3.17.0

# =============================================================================
# LLM Integration
# =============================================================================
anthropic>=0.18.0
openai>=1.10.0

# =============================================================================
# Web Framework (Admin Dashboard)
# =============================================================================
fastapi>=0.109.0
uvicorn[standard]>=0.25.0
python-multipart>=0.0.6
jinja2>=3.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# =============================================================================
# Scheduling
# =============================================================================
apscheduler>=3.10.0

# =============================================================================
# Utilities
# =============================================================================
python-slugify>=8.0.0
python-dotenv>=1.0.0
PyYAML>=6.0.0
markdown>=3.5.0

# =============================================================================
# Development & Testing
# =============================================================================
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0  # Also used for async test client

# Code Quality
black>=23.12.0
ruff>=0.1.0
mypy>=1.8.0

# Type Stubs
types-PyYAML>=6.0.0
types-python-slugify>=8.0.0
types-beautifulsoup4>=4.12.0
types-passlib>=1.7.0

# =============================================================================
# Production
# =============================================================================
gunicorn>=21.2.0
